extract dashboard results out into separate hook
